+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 # Added
 
+- [TD-4535] Structures domains CSV upload bulk update
 - [TD-3233] Rule result remediation plan
 - [TD-4271] Support for linking implementations with business concepts
 - [TD-4621] Include data structure version `name` in GraphQL schema